Added several eye tracking methods
I experimented with several different methods for isolating the eye or
certain parts of the eye in the image.

Unsuccessful methods were: Canny edge detection (this picks up the
projector screen reflection well), Haar cascade eye detection (too
unreliable with default trained classifier), Sobel filtering (too much
noise), and different types of image value thresholding.

The most successful method, which I am still working on, is converting
to HSV color space, thresholding for the hue value of the white part of
the eye, and then using a median blue to eliminate noise around the
eye. It’s pretty good in consistently detecting the white part of the
eye, but it also picks some places outside the eye. I am working to
refine this approach.